fix(positioning): fix errors on invalid position
